Fix printf format so that icc on Linux accepts it. An unsigned short argument should get a %d format, since it gets promoted to int, not unsigned int, according to the C standard.

+++ gdk/gdk_bbp.c
@@ -1222,7 +1222,7 @@ BBPexit(void)
 static inline int
 heap_entry(FILE *fp, COLrec *col)
 {
-	return fprintf(fp, " %s %u %u %u " BUNFMT " " BUNFMT " " BUNFMT " "
+	return fprintf(fp, " %s %d %d %d " BUNFMT " " BUNFMT " " BUNFMT " "
 		       BUNFMT " " OIDFMT " " OIDFMT " " SZFMT " " SZFMT " %d",
 		       col->type >= 0 ? BATatoms[col->type].name : ATOMunknown_name(col->type),
 		       col->width,
@@ -1276,8 +1276,10 @@ new_bbpentry(FILE *fp, bat i)
 	}
 #endif
 
-	if (fprintf(fp, SSZFMT " %d %s %s %s %d %u " BUNFMT " " BUNFMT " " BUNFMT " " BUNFMT " " BUNFMT " %u %u %u %u",	/* BAT info */
-		    (ssize_t) i, BBP_status(i) & BBPPERSISTENT,
+	if (fprintf(fp, SSZFMT " %d %s %s %s %d %d " BUNFMT " " BUNFMT " "
+		    BUNFMT " " BUNFMT " " BUNFMT " %d %d %d %d", /* BAT info */
+		    (ssize_t) i,
+		    BBP_status(i) & BBPPERSISTENT,
 		    BBP_logical(i),
 		    BBP_logical(-i) ? BBP_logical(-i) : BBPNONAME,
 		    BBP_physical(i),
@@ -2837,11 +2839,11 @@ BBPtrim_select(size_t target, int dirty)
 				PTRFMTCAST(void *)b);
 
 			fprintf(stderr,
-				"#            (cnt=" BUNFMT ", mode=%u, "
+				"#            (cnt=" BUNFMT ", mode=%d, "
 				"refs=%d, wait=%d, parent=%d,%d, "
 				"lastused=%d,%d,%d)\n",
 				bbptrim[cur].cnt,
-				b->batPersistence,
+				(int) b->batPersistence,
 				BBP_refs(b->batCacheid),
 				(BBP_status(b->batCacheid) & BBPWAITING) != 0,
 				VIEWhparent(b),
